// web/src/api/generated/models/add-todo-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Template FastAPI React
 *  ### Description A RESTful API for handling todos items.  Anyone in Equinor are authorized to run these calculations.  * Click **Authorize** to login and start testing.  ### Resources  * [Docs](https://equinor.github.io/template-fastapi-react/)  * [Github](https://github.com/equinor/template-fastapi-react)   For questions about usage or expanding the service with new simulations, create issue on Github.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AddTodoResponse
 */
export interface AddTodoResponse {
    /**
     * 
     * @type {string}
     * @memberof AddTodoResponse
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof AddTodoResponse
     */
    title: string;
    /**
     * 
     * @type {boolean}
     * @memberof AddTodoResponse
     */
    is_completed?: boolean;
}